Fixed circular dependency issue
