--- src/cleanvision/dataset/__init__.py
@@ -1,1 +1,0 @@
-Ï
